+++ packages/quick_actions/quick_actions_ios/lib/quick_actions_ios.dart
@@ -11,9 +11,9 @@ export 'package:quick_actions_platform_interface/types/types.dart';
 const MethodChannel _channel =
     MethodChannel('plugins.flutter.io/quick_actions_ios');
 
-/// An implementation of [QuickActionsPlatform] that uses method channels.
+/// An implementation of [QuickActionsPlatform] that for iOS.
 class QuickActionsIos extends QuickActionsPlatform {
-  /// Registers this class as the default instance of [PathProviderPlatform].
+  /// Registers this class as the default instance of [QuickActionsPlatform].
   static void registerWith() {
     QuickActionsPlatform.instance = QuickActionsIos();
   }