+++ packages/quick_actions/quick_actions_ios/test/quick_actions_ios_test.dart
@@ -7,6 +7,7 @@ import 'dart:async';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:quick_actions_ios/quick_actions_ios.dart';
+import 'package:quick_actions_platform_interface/quick_actions_platform_interface.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
@@ -24,6 +25,11 @@ void main() {
       });
 
       log.clear();
+    });
+
+    test('registerWith() registers correct instance', () {
+      QuickActionsIos.registerWith();
+      expect(QuickActionsPlatform.instance, isA<QuickActionsIos>());
     });
 
     group('#initialize', () {