+++ packages/quick_actions/quick_actions_ios/CHANGELOG.md
@@ -1,4 +1,4 @@
 
 ## 0.6.0+9
 
-* Extract iOS implementation from quick_actions.
+* Switches to a package-internal implementation of the platform interface.